Cart view working from every activity
