+++ app/src/main/java/com/example/nermingraca/bitbayapp/activities/CartActivity.java
@@ -12,11 +12,17 @@ import com.example.nermingraca.bitbayapp.CustomListAdapter;
 import com.example.nermingraca.bitbayapp.R;
 import com.example.nermingraca.bitbayapp.models.Product;
 import com.example.nermingraca.bitbayapp.models.User;
+import com.example.nermingraca.bitbayapp.service.ServiceRequest;
+import com.example.nermingraca.bitbayapp.singletons.UserData;
+import com.squareup.okhttp.Callback;
+import com.squareup.okhttp.Request;
+import com.squareup.okhttp.Response;
 
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -32,7 +38,7 @@ public class CartActivity extends ActionBarActivity {
 
         Intent intent = getIntent();
         String json = intent.getStringExtra("jsonProducts");
-        //Log.d("DEBUG", json);
+        Log.d("DEBUG in Cart Activity", json);
         List<Product> products = productsFromJson(json);
 
         mProductList = (ListView)findViewById(R.id.cart_list);
@@ -118,11 +124,47 @@ public class CartActivity extends ActionBarActivity {
         }
 
         if (id == R.id.cart_action) {
-            Intent intent = new Intent(CartActivity.this, CartActivity.class);
-            startActivity(intent);
+            toCart();
             return true;
         }
 
         return super.onOptionsItemSelected(item);
+    }
+
+    private void toCart() {
+        int buyerId = UserData.getInstance().getId();
+        String url = getString(R.string.service_get_cart);
+        JSONObject json = new JSONObject();
+
+        try {
+            json.put("userId", buyerId);
+        } catch (JSONException e) {
+            e.printStackTrace();
+            Log.e("ERROR", e.getMessage());
+        }
+        String jsonString = json.toString();
+        Log.d("DEBUG", jsonString);
+        Callback callback = response();
+        ServiceRequest.post(url, jsonString, callback);
+    }
+
+    public Callback response() {
+        return new Callback() {
+            @Override
+            public void onFailure(Request request, IOException e) {
+                Log.e("ERROR", e.getMessage());
+            }
+
+            @Override
+            public void onResponse(Response response) throws IOException {
+
+                String responseJson = response.body().string();
+                Log.d("DEBUG", responseJson);
+                Intent toCart = new Intent(CartActivity.this, CartActivity.class);
+                toCart.putExtra("jsonProducts", responseJson);
+                startActivity(toCart);
+            }
+        };
+
     }
 }
