+++ app/src/main/java/com/example/nermingraca/bitbayapp/activities/ProfileActivity.java
@@ -3,6 +3,7 @@ package com.example.nermingraca.bitbayapp.activities;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.ActionBarActivity;
+import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
@@ -14,11 +15,19 @@ import com.example.nermingraca.bitbayapp.CustomListAdapter;
 import com.example.nermingraca.bitbayapp.R;
 import com.example.nermingraca.bitbayapp.models.Product;
 import com.example.nermingraca.bitbayapp.models.User;
+import com.example.nermingraca.bitbayapp.service.ServiceRequest;
 import com.example.nermingraca.bitbayapp.singletons.AppController;
 import com.example.nermingraca.bitbayapp.singletons.ProductFeed;
 import com.example.nermingraca.bitbayapp.singletons.UserData;
 import com.example.nermingraca.bitbayapp.singletons.UserFeed;
+import com.squareup.okhttp.Callback;
+import com.squareup.okhttp.Request;
+import com.squareup.okhttp.Response;
 
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
@@ -132,11 +141,47 @@ public class ProfileActivity extends ActionBarActivity {
         }
 
         if (id == R.id.cart_action) {
-            Intent intent = new Intent(ProfileActivity.this, CartActivity.class);
-            startActivity(intent);
+            toCart();
             return true;
         }
 
         return super.onOptionsItemSelected(item);
+    }
+
+    private void toCart() {
+        int buyerId = UserData.getInstance().getId();
+        String url = getString(R.string.service_get_cart);
+        JSONObject json = new JSONObject();
+
+        try {
+            json.put("userId", buyerId);
+        } catch (JSONException e) {
+            e.printStackTrace();
+            Log.e("ERROR", e.getMessage());
+        }
+        String jsonString = json.toString();
+        Log.d("DEBUG", jsonString);
+        Callback callback = response();
+        ServiceRequest.post(url, jsonString, callback);
+    }
+
+    public Callback response() {
+        return new Callback() {
+            @Override
+            public void onFailure(Request request, IOException e) {
+                Log.e("ERROR", e.getMessage());
+            }
+
+            @Override
+            public void onResponse(Response response) throws IOException {
+
+                String responseJson = response.body().string();
+                Log.d("DEBUG", responseJson);
+                Intent toCart = new Intent(ProfileActivity.this, CartActivity.class);
+                toCart.putExtra("jsonProducts", responseJson);
+                startActivity(toCart);
+            }
+        };
+
     }
 }
